Remove the superseded v5/v6/v7 search schema revisions
v8 is self-contained and is the only revision production creates indexes
from, so the older chained revisions (v7 -> v6 -> v5) are no longer
referenced and can be deleted. This removes their modules, drops them from
the revision directory, and repoints the remaining test/mock usages at v8.

The v8 module test absorbs the author char-filter test that previously
lived against v5 in test_external_search, and drops its v7-equivalence
test now that v7 is gone.
