+++ src/palace/manager/search/revision_directory.py
@@ -2,12 +2,9 @@ from collections.abc import Mapping
 
 from palace.manager.core.config import CannotLoadConfiguration
 from palace.manager.search.revision import SearchSchemaRevision
-from palace.manager.search.v5 import SearchV5
-from palace.manager.search.v6 import SearchV6
-from palace.manager.search.v7 import SearchV7
 from palace.manager.search.v8 import SearchV8
 
-REVISIONS = [SearchV5(), SearchV6(), SearchV7(), SearchV8()]
+REVISIONS = [SearchV8()]
 
 
 class SearchRevisionDirectory:
